# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: Clothing Line Commerce API
  description: |
    Headless e-commerce API for the clothing line storefront and admin consoles.

    This contract is the single source of truth shared by the storefront (Next.js), the
    backend adapters, and the mock (Prism). It is organised into six bounded contexts:

      - Admin     — catalogue administration, sales channels, regional pricing, tax,
                    promotions, roles, bulk imports, queue (dead-letter) operations.
      - Catalog   — storefront catalogue reads, search, category trees, cross-selling,
                    variant availability and product reviews.
      - Cart      — cart session lifecycle, line items (variant + custom B2B), discounts,
                    guest-cart merging.
      - Checkout  — shipping addresses, logistics quotes, insurance quotes, payment
                    session initialisation and async payment webhooks.
      - Customers — registration, JWT authentication, session revocation, password reset,
                    address books, B2B business units and quote requests.
      - Logistics — order returns (RMA), swaps, order edits, fulfilment dispatch, draft
                    orders and courier tracking webhooks.

    Conventions:

      * Money is always expressed as a non-negative **integer in minor units**
        (Kobo / Cents). Monetary fields end with the suffix `Minor` (e.g. `amountMinor`,
...
  version: 1.0.0
servers:
  - url: http://localhost:4000
...
security:
  - bearerAuth: []
paths:
...
  /store/carts:
    post:
      tags: [Cart]
      summary: Initialize a new cart session.
      description: |
        Creates a transient cart bound to a region and sales channel. Both identifiers are
        required. An optional email and ISO country code may be attached.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitializeCartRequest"
      responses:
        "200":
          description: Newly created cart.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "400":
          $ref: "#/components/responses/StandardErrorResponse"
        "404":
          $ref: "#/components/responses/StandardErrorResponse"
        "500":
          $ref: "#/components/responses/StandardErrorResponse"
  /store/carts/{id}:
    get:
      tags: [Cart]
      summary: Retrieve the current state of a cart.
      security: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Cart DTO.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
        "404":
          $ref: "#/components/responses/StandardErrorResponse"
        "500":
          $ref: "#/components/responses/StandardErrorResponse"
  /store/carts/{id}/line-items:
    post:
      tags: [Cart]
      summary: Add a product variant line item to the cart.
      description: |
        Resolves the regional price server-side. Rejects when inventory is insufficient and
        backordering is disabled (`OUT_OF_STOCK`) or when no regional price exists
        (`REGIONAL_PRICE_MISSING`).
      security: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLineItemRequest"
      responses:
        "204":
          description: Line item added. Retrieve the cart via GET /store/carts/{id}.
        "400":
          $ref: "#/components/responses/StandardErrorResponse"
        "404":
          $ref: "#/components/responses/StandardErrorResponse"
        "409":
          $ref: "#/components/responses/StandardErrorResponse"
        "500":
          $ref: "#/components/responses/StandardErrorResponse"
  /store/carts/{id}/line-items/custom:
    post:
      tags: [Cart]
      summary: Add a custom (B2B) line item with an explicit title and unit price.
      description: |
        Custom line items have no variant; the client supplies `title`, `quantity` and
        `unitPriceMinor` (integer minor units, non-negative).
      security: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCustomLineItemRequest"
      responses:
        "204":
          description: Custom line item added.
        "400":
          $ref: "#/components/responses/StandardErrorResponse"
        "404":
          $ref: "#/components/responses/StandardErrorResponse"
        "500":
          $ref: "#/components/responses/StandardErrorResponse"
  /store/carts/{id}/line-items/{line_id}:
    put:
      tags: [Cart]
      summary: Update the quantity of an existing line item.
      description: |
        Variant-backed items are re-validated against current inventory; custom items only
        respect the quantity bounds.
      security: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: line_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLineItemQuantityRequest"
      responses:
        "204":
          description: Quantity updated.
        "400":
          $ref: "#/components/responses/StandardErrorResponse"
        "404":
          $ref: "#/components/responses/StandardErrorResponse"
        "409":
          $ref: "#/components/responses/StandardErrorResponse"
        "500":
          $ref: "#/components/responses/StandardErrorResponse"
    delete:
      tags: [Cart]
      summary: Remove a line item from the cart.
      security: []
      parameters:
        - $ref: "#/components/parameters/PathId"
        - name: line_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Line item removed.
        "400":
          $ref: "#/components/responses/StandardErrorResponse"
        "404":
          $ref: "#/components/responses/StandardErrorResponse"
        "500":
          $ref: "#/components/responses/StandardErrorResponse"
  /store/carts/{id}/discount:
    post:
      tags: [Cart]
      summary: Apply a promotion code to the cart.
      description: |
        Only one active discount may be applied per cart. The code is case-insensitive and
        normalised to uppercase. Rejects inactive/unknown codes and codes whose minimum spend
        is not met.
      security: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyDiscountRequest"
      responses:
        "204":
          description: Discount applied.
        "400":
          $ref: "#/components/responses/StandardErrorResponse"
        "404":
          $ref: "#/components/responses/StandardErrorResponse"
        "409":
          $ref: "#/components/responses/StandardErrorResponse"
        "500":
          $ref: "#/components/responses/StandardErrorResponse"
  /store/carts/{id}/merge:
    post:
      tags: [Cart]
      summary: Merge a guest cart into the authenticated customer's active cart.
      description: |
        Matching line items (same variant, unit price, title and metadata) are merged by
        summing quantities; unmatched items are copied. The guest cart is deleted.
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeGuestCartRequest"
      responses:
        "204":
          description: Carts merged.
        "400":
          $ref: "#/components/responses/StandardErrorResponse"
        "404":
          $ref: "#/components/responses/StandardErrorResponse"
        "409":
          $ref: "#/components/responses/StandardErrorResponse"
        "500":
          $ref: "#/components/responses/StandardErrorResponse"

  # =====================================================================
  # CHECKOUT
  # =====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by POST /store/auth.
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    StandardErrorResponse:
      description: Standardized error envelope used by every non-2xx response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StandardError"
  schemas:
...
    StandardError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                # Mirrors the ErrorCode union in DomainError.ts exactly (one
                # entry per code, no duplicates, no stale codes). Keep this list
                # in sync with the domain union + mapDomainErrorToHttp.
                - VALIDATION_ERROR
                - INVALID_EMAIL
                - NEGATIVE_AMOUNT
                - INVALID_CURRENCY
                - INVALID_OPERATION
                - INVALID_STATE
                - INVALID_STATUS_TRANSITION
                - OUT_OF_STOCK
                - REGIONAL_PRICE_MISSING
                - INTERNAL_ERROR
                - JOB_PROCESSING_ERROR
                - PAYMENT_VERIFICATION_FAILED
                - LOGISTICS_VERIFICATION_FAILED
                - LOGISTICS_EVENT_FULFILLMENT_NOT_FOUND
                - EXTERNAL_SERVICE_TIMEOUT
                - EXTERNAL_SERVICE_UNAVAILABLE
                - EXTERNAL_SERVICE_ERROR
                - LOCK_ACQUISITION_FAILED
                - UNSUPPORTED_OPERATION
                - ORDER_ALREADY_FULFILLED
                - INVALID_RETURN_QUANTITY
                - DUPLICATE_DRAFT_ORDER
                - PAYMENT_REQUIRED
                - INVALID_RETURN_ITEM
                - INVALID_INPUT
                - REGION_NOT_FOUND
                - DUPLICATE_TRANSACTION
                - TRANSACTION_NOT_FOUND
                - INVALID_PAYMENT_AMOUNT
                - PERMISSION_DENIED
                - DUPLICATE_QUOTE
                - UNAUTHORIZED
                - UNAUTHORIZED_REVIEW
                - INVALID_CREDENTIALS
                - UNAUTHORIZED_ACCESS
                - CUSTOMER_ALREADY_EXISTS
                - COMPLIANCE_VIOLATION
                - ACCOUNT_DISABLED
                - ACCOUNT_LOCKED
                - BUSINESS_UNIT_ALREADY_EXISTS
                - PAYMENT_DECLINED
                - REFUND_REQUIRES_REVIEW
                - SHIPMENT_REQUIRES_RECONCILIATION
                - INVALID_SIGNATURE
                - RESOURCE_NOT_FOUND
                - CART_NOT_FOUND
                - PRODUCT_NOT_FOUND
                - INSUFFICIENT_INVENTORY
                - INSUFFICIENT_SINGLE_LOCATION_STOCK
                - SOURCING_FAILED
              example: OUT_OF_STOCK
            message:
              type: string
              example: Requested quantity exceeds available inventory.
            details:
              type: object
              additionalProperties: true

    # ------------------------------------------------------------------
    # Catalog DTOs
    # ------------------------------------------------------------------
...
    Cart:
      type: object
      description: Cart session DTO including computed totals and applied promotion.
      required: [id, regionId, salesChannelId, status, items, cartTotalMinor]
      properties:
        id:
          type: string
          format: uuid
        regionId:
          type: string
          format: uuid
        salesChannelId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        createdAt:
          type: string
          format: date-time
        countryCode:
          type: string
          nullable: true
          example: NG
        currency:
          type: string
          example: NGN
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        taxAmountMinor:
          type: integer
          minimum: 0
          nullable: true
          description: Applied tax in minor units; null before tax is calculated.
        metadata:
          type: object
          additionalProperties: true
          description: Arbitrary JSONB metadata attached to the cart.
        frozen:
          type: boolean
        frozenReason:
          type: string
          nullable: true
        frozenAt:
          type: string
          format: date-time
          nullable: true
        orderId:
          type: string
          format: uuid
          nullable: true
        convertedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [active, converted]
        paymentStatus:
          type: string
          enum: [pending, initialized, paid]
        paymentInitialized:
          type: boolean
        paymentAuthorizationUrl:
          type: string
          nullable: true
        paymentInitializedAt:
          type: string
          format: date-time
          nullable: true
        cartTotalMinor:
          type: integer
          minimum: 0
          description: Sum of all line-item totals in minor units.
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartLineItem"
        appliedPromotion:
          $ref: "#/components/schemas/AppliedPromotion"
    CartLineItem:
      type: object
      description: A single cart line item. Variant items reference a variant; custom (B2B) items carry a title and client-supplied unit price instead.
      required: [id, cartId, quantity, unitPriceMinor, createdAt]
      properties:
        id:
          type: string
          format: uuid
        cartId:
          type: string
          format: uuid
        variantId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          nullable: true
          description: Required for custom items.
        quantity:
          type: integer
          minimum: 1
          maximum: 1000000
        unitPriceMinor:
          type: integer
          minimum: 0
          description: Unit price in minor units.
        lineTotalMinor:
          type: integer
          minimum: 0
          description: Computed unitPriceMinor * quantity in minor units.
        metadata:
          type: object
          additionalProperties: true
          description: Arbitrary JSONB metadata attached to the line item.
        createdAt:
          type: string
          format: date-time
    AppliedPromotion:
      type: object
      description: Promotion projection attached to a cart (only id and code are exposed).
      required: [id, code]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string

    # ------------------------------------------------------------------
    # Checkout DTOs
    # ------------------------------------------------------------------
    ShippingAddress:
      type: object
      description: Checkout shipping address. Stored as JSONB; unknown keys are preserved.
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        company:
          type: string
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        countryCode:
          type: string
          example: NG
        isBusiness:
          type: boolean
          default: false
        metadata:
          type: object
          additionalProperties: true
      additionalProperties: true
...
    InitializeCartRequest:
      type: object
      required: [regionId, salesChannelId]
      properties:
        regionId:
          type: string
          format: uuid
        salesChannelId:
          type: string
          format: uuid
        email:
          type: string
          format: email
          maxLength: 254
        countryCode:
          type: string
          pattern: "^[A-Z]{2}$"
          description: ISO 3166-1 alpha-2 country code, uppercased.
    AddLineItemRequest:
      type: object
      required: [variantId, quantity]
      properties:
        variantId:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 1000000
        metadata:
          type: object
          additionalProperties: true
    AddCustomLineItemRequest:
      type: object
      required: [title, quantity, unitPriceMinor]
      properties:
        title:
          type: string
          maxLength: 1000
        quantity:
          type: integer
          minimum: 1
          maximum: 1000000
        unitPriceMinor:
          type: integer
          minimum: 0
          description: Custom unit price in minor units.
    UpdateLineItemQuantityRequest:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer
          minimum: 1
          maximum: 1000000
    ApplyDiscountRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          pattern: "^[A-Z0-9-_]+$"
          description: Promotion code; case-insensitive and uppercased server-side.
    MergeGuestCartRequest:
      type: object
      required: [guestCartId, customerId]
      properties:
        guestCartId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
...